LongFormatFileList/ShortFormatFileListをFileListに統合
